+++ filer/models/__init__.py
@@ -3,4 +3,5 @@ from filer.models.clipboardmodels import *
 from filer.models.filemodels import *
 from filer.models.foldermodels import *
 from filer.models.imagemodels import *
+from filer.models.videomodels import *
 from filer.models.virtualitems import *
